feat: adapt to ai-foundation API changes and improve AI extraction UI
- Replace api-1.0.0-SNAPSHOT.jar with api-0.0.0-7e636df07435.jar
- Update AiModelService::defaultLanguageModel to ::languageModel
- Make ai-foundation an optional dependency (ai-foundation?: *)
- Split LinkAiEndpoint into LinkAiEndpoint (recent-comments) and
  LinkAiExtractEndpoint (ai-extract) to avoid class loading failures
  when ai-foundation is not installed
- Add rssUrl field to LinkCommentAnalysisResult for RSS feed extraction
- Update prompts to include RSS URL extraction
- Improve AI extraction UI: card-style container, relative time display,
  loading/empty states, better comment list styling
- Use utils.date.timeAgo from @halo-dev/ui-shared for time formatting
- Apply extracted rssUrl to RSS feed URLs automatically
