+++ src/main/resources/plugin.yaml
@@ -7,7 +7,7 @@ metadata:
 spec:
   requires: ">=2.22.5"
   pluginDependencies:
-    ai-foundation: "*"
+    ai-foundation?: "*"
   author:
     name: Halo
     website: https://github.com/halo-dev
